+++ src/imaging_source_recorder/gui.py
