+++ src/imaging_source_recorder/imaging_source_recorder.py
