+++ src/imaging_source_recorder/resourceselector.py
