+++ src/imaging_source_recorder/video_recorder_interface.py
